Inline NoError
