+++ test/test-runner/src/main.rs
@@ -1,5 +1,4 @@
 mod index;
-mod infallible;
 mod phases;
 mod runner;
 mod suites;
